support envfile in golang and rr binary
